classes: cogip-genimage: test bmap file before creating symlink
From Yocto manual:
  If do_deploy is not already in the shared state cache or if its
  input checksum (signature) has changed from when the output was
  cached, the task runs to populate the shared state cache, after
  which the contents of the shared state cache is copied to
  ${DEPLOY_DIR_IMAGE}.

So if for any reason do_genimage has been cached and not do_deploy,
the GENIMAGE_IMAGE_NAME variable (containing the datetime) would
have change, leading to link creation failure.

[1] https://www.yoctoproject.org/docs/latest/overview-manual/overview-manual.html#shared-state

+++ classes/cogip-genimage.bbclass
@@ -11,7 +11,9 @@ do_genimage[depends] += " \
 
 # Link to bmap file
 do_deploy_append() {
-    ln -sf ${GENIMAGE_IMAGE_NAME}.${GENIMAGE_IMAGE_SUFFIX}.bmap ${DEPLOYDIR}/${GENIMAGE_IMAGE_LINK_NAME}.${GENIMAGE_IMAGE_SUFFIX}.bmap
+    if [ -f ${DEPLOYDIR}/${GENIMAGE_IMAGE_NAME}.${GENIMAGE_IMAGE_SUFFIX}.bmap ]; then
+        ln -sf ${GENIMAGE_IMAGE_NAME}.${GENIMAGE_IMAGE_SUFFIX}.bmap ${DEPLOYDIR}/${GENIMAGE_IMAGE_LINK_NAME}.${GENIMAGE_IMAGE_SUFFIX}.bmap
+    fi
 }
 
 # Link rootfs image to genimage image name
